refactor(benchmark): 목업 데이터를 fixtures/로 공통화
- 시나리오별 목업 데이터를 benchmarks/fixtures/로 분리
- CONV_STYLES도 fixtures로 공통화
- v1/v2 runner에서 공유 fixtures 참조하도록 변경
