+++ benchmarks/tool-calling/v1/scenarios.ts
@@ -5,53 +5,11 @@
 import {
   MOCK_DOCUMENTS,
   MOCK_CAREER_NOTES,
-} from "../../chat-pipeline/v1/common"
+  CONV_STYLES,
+} from "../../fixtures/mock-data"
 
 // polite 스타일의 앞 4쌍 (8개 메시지) — 공통 대화 히스토리
-export const BASE_CONVERSATION: Array<{
-  role: "user" | "assistant"
-  content: string
-}> = [
-  {
-    role: "user",
-    content: "안녕하세요, 네이버 클라우드 지원하려고 합니다.",
-  },
-  {
-    role: "assistant",
-    content:
-      "네, 네이버 클라우드 시니어 백엔드 개발자 포지션에 지원하시는군요! 채용공고와 참고자료를 확인해보겠습니다. 어떤 부분의 자기소개서를 먼저 작성할까요?",
-  },
-  {
-    role: "user",
-    content:
-      "일단 전체적인 구조를 잡아주세요. 어떤 항목을 넣으면 좋을지 추천해주세요.",
-  },
-  {
-    role: "assistant",
-    content:
-      "네이버 클라우드 채용공고를 보면 Kubernetes, MSA, 대규모 분산 시스템, 오픈소스 기여를 중시합니다. 다음 구조를 추천드립니다:\n\n1. 지원 동기\n2. 핵심 역량\n3. 프로젝트 경험\n4. 성장 가능성\n\n어떤 항목부터 시작할까요?",
-  },
-  {
-    role: "user",
-    content:
-      "좋네요. 2번 핵심 역량부터 시작하죠. 제 Kubernetes 경험을 중심으로 써주세요.",
-  },
-  {
-    role: "assistant",
-    content:
-      "Kubernetes 관련 경험을 정리해보겠습니다. deploy-ez에서 Docker + Kubernetes 자동 배포 경험, 경력기술서에 ArgoCD 기반 GitOps 파이프라인 경험이 있으시네요.",
-  },
-  {
-    role: "user",
-    content:
-      "네, deploy-ez 프로젝트에서의 Kubernetes 경험과 회사에서의 경험을 합쳐서 하나의 스토리로 만들어주세요.",
-  },
-  {
-    role: "assistant",
-    content:
-      "좋습니다. deploy-ez 오픈소스에서 시작된 K8s 자동화가 회사에서 12개 서비스 GitOps로 이어진 스토리로 작성하겠습니다.",
-  },
-]
+export const BASE_CONVERSATION = CONV_STYLES.polite.slice(0, 8)
 
 // ---------------------------------------------------------------------------
 // 시나리오 정의
